requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
ansicon==1.89.0
anyio==3.6.2
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
croniter==1.3.14
dateutils==0.6.12
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
entrypoints==0.4
executing==1.2.0
fastapi==0.88.0
filelock==3.12.0
flatbuffers==23.5.9
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
inquirer==3.1.3
ipykernel==6.20.2
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.1.4
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
onnx==1.14.0
onnxruntime==1.14.1
ordered-set==4.1.0
packaging==23.0
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==12.0.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.14.0
PyJWT==2.7.0
Pympler==1.0.1
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
#pywin32==305
PyYAML==6.0
pyzmq==25.0.0
readchar==4.0.5
regex==2023.5.5
requests==2.30.0
rich==13.3.5
safetensors==0.3.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
streamlit==1.22.0
sympy==1.12
tenacity==8.2.2
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.8.1
transformers==4.29.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0